backend/requirements.txt
django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
pillow==12.2.0
django-cors-headers==4.9.0
